requirements.txt
flask>=3.0.0
requests>=2.31.0
python-dotenv>=1.0.1
python-pptx>=0.6.23
